Revert "Audio player tidying but without breaking things (#16402)"
This reverts commit d20d00daadea73b9b19a6ccfcada9f4cb1af7b12.
