+++ dotcom-rendering/src/layouts/AudioLayout.tsx
@@ -9,6 +9,7 @@ import { StraightLines } from '@guardian/source-development-kitchen/react-compon
 import { AdPortals } from '../components/AdPortals.island';
 import { AdSlot, MobileStickyContainer } from '../components/AdSlot.web';
 import { AffiliateDisclaimer } from '../components/AffiliateDisclaimer';
+import { AppsAudioPlayer } from '../components/AppsAudioPlayer.island';
 import { AppsFooter } from '../components/AppsFooter.island';
 import { ArticleBody } from '../components/ArticleBody';
 import { ArticleContainer } from '../components/ArticleContainer';
@@ -16,7 +17,7 @@ import { ArticleHeadline } from '../components/ArticleHeadline';
 import { ArticleMetaApps } from '../components/ArticleMeta.apps';
 import { ArticleMeta } from '../components/ArticleMeta.web';
 import { ArticleTitle } from '../components/ArticleTitle';
-import { AudioPlayer } from '../components/AudioPlayer/AudioPlayer';
+import { AudioPlayerWrapper } from '../components/AudioPlayerWrapper.island';
 import { Border } from '../components/Border';
 import { Carousel } from '../components/Carousel.island';
 import { DirectoryPageNavIsland } from '../components/DirectoryPageNavIsland';
@@ -26,6 +27,7 @@ import { GridItem } from '../components/GridItem';
 import { HeaderAdSlot } from '../components/HeaderAdSlot';
 import { Island } from '../components/Island';
 import { LabsHeader } from '../components/LabsHeader';
+import { formatAudioDuration } from '../components/ListenToArticle.island';
 import { Masthead } from '../components/Masthead/Masthead';
 import { MostViewedFooterData } from '../components/MostViewedFooterData.island';
 import { MostViewedFooterLayout } from '../components/MostViewedFooterLayout';
@@ -329,18 +331,37 @@ export const AudioLayout = (props: WebProps | AppProps) => {
 							</div>
 						</GridItem>
 						<GridItem area="media">
-							{audioData && (
+							{isWeb && audioData && (
 								<Island
 									priority="critical"
 									defer={{ until: 'visible' }}
 								>
-									<AudioPlayer
-										audioData={audioData}
-										isSensitive={article.config.isSensitive}
+									<AudioPlayerWrapper
+										contentIsNotSensitive={
+											!article.config.isSensitive
+										}
 										isAcastEnabled={
 											!!article.config.switches.acast
 										}
-										isApps={isApps}
+										src={audioData.audioDownloadUrl}
+										mediaId={audioData.mediaId}
+									/>
+								</Island>
+							)}
+							{isApps && audioData && (
+								<Island
+									priority="critical"
+									defer={{ until: 'visible' }}
+								>
+									<AppsAudioPlayer
+										audioDuration={
+											typeof audioData.durationSeconds ===
+											'number'
+												? formatAudioDuration(
+														audioData.durationSeconds,
+													)
+												: undefined
+										}
 									/>
 								</Island>
 							)}
